@@ -1,9 +1,6 @@
 source 'https://rubygems.org'
 
-ruby '1.9.3'
-#ruby-gemset=railstutorial_rails_4_0
-
-gem 'rails', '3.2.13'
+gem 'rails', '4.0.0'
 gem 'bootstrap-sass', '2.3.2.0'
 gem 'bcrypt-ruby', '3.0.1'
 gem 'will_paginate', '3.0.4'
@@ -25,13 +22,12 @@ end
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+	gem 'sass-rails'
+	gem 'uglifier', '2.1.1'
+	gem 'coffee-rails'
 
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   # gem 'therubyracer', :platforms => :ruby
-
-  gem 'uglifier', '>= 1.0.3'
 end
 
 gem 'jquery-rails'
