@@ -35,7 +35,7 @@ class UsersController < ApplicationController
   end
 
 	def create
-   	@user = User.new(params[:user])
+   	@user = User.new(user_params)
    	if @user.save
       sign_in @user
    		flash[:success] = "Welcome to Jynx!"
