@@ -10,7 +10,7 @@
 #
 
 class User < ActiveRecord::Base
-  attr_accessible :name, :email, :password, :password_confirmation
+  # attr_accessible :name, :email, :password, :password_confirmation
   has_secure_password
 
  before_save { |user| user.email = email.downcase }
